right after 'Quick Setup' finished https://github.com/caoxiemeihao/nuxt-electron#quick-setup
note that I have downgraded `electron` from 38.0.0 to 37.4.0 purposely, otherwise build fails as follows:
Could not detect abi for version 38.0.0 and runtime electron.  Updating "node-abi" might help solve this issue if it is a new release of electron  failedTask=build stackTrace=Error: Could not detect abi for version 38.0.0 and runtime electron.  Updating "node-abi" might help solve this issue if it is a new release of electron.

latest node-abi@4.13.0 might work, but I don't want to 'manually' resolve package versions.
also 37.4.0 is known to work, where as 38.0.0 is not.